+++ src/features/campaign/components/TimingForm.tsx
@@ -1,17 +1,17 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import React, { FormEventHandler, useContext, useEffect, useState } from 'react';
-import { CampaignContext } from '../../features/campaign/context/CampaignContext';
-import Dropdown from '../../shared/components/ui/Dropdown';
+import { CampaignContext } from '../context/CampaignContext';
+import Dropdown from '../../../shared/components/ui/Dropdown';
 
 import dayjs, { Dayjs } from 'dayjs';
 import { DatePicker } from '@mui/x-date-pickers/DatePicker';
 import { LocalizationProvider } from '@mui/x-date-pickers/LocalizationProvider';
 import { AdapterDayjs } from '@mui/x-date-pickers/AdapterDayjs';
-import FormsContainer from '../../shared/components/layout/FormsContainer';
-import BackButton from '../../shared/components/buttons/BackButton';
-import NextButton from '../../shared/components/buttons/NextButton';
-import posthog from '../../shared/utils/posthog';
-import { useAuth } from '../../features/auth/context/AuthContext';
+import FormsContainer from '../../../shared/components/layout/FormsContainer';
+import BackButton from '../../../shared/components/buttons/BackButton';
+import NextButton from '../../../shared/components/buttons/NextButton';
+import posthog from '../../../shared/utils/posthog';
+import { useAuth } from '../../auth/context/AuthContext';
 
 interface TimingFormProps {
   handleNext?: any;
