+++ src/features/campaign/pages/Campaign.tsx
@@ -3,11 +3,11 @@ import { useLocation, useNavigate } from 'react-router-dom';
 import CampaignStepsBar from '../components/CampaignStepsBar';
 import CustomComponent from '../components/CustomComponent';
 import ErrorToast from '../../../shared/components/feedback/ErrorToast';
-import AudienceForm from '../../../components/Forms/AudienceForm';
-import CampaignBriefForm from '../../../components/Forms/CampaignBriefForm';
-import PurposeForm from '../../../components/Forms/PurposeForm';
+import AudienceForm from '../components/AudienceForm';
+import CampaignBriefForm from '../components/CampaignBriefForm';
+import PurposeForm from '../components/PurposeForm';
 import ServiceForm from '../../onboarding/components/ServiceForm';
-import TimingForm from '../../../components/Forms/TimingForm';
+import TimingForm from '../components/TimingForm';
 import Sidebar from '../../../shared/components/layout/Sidebar';
 import { useAuth } from '../../auth/context/AuthContext';
 import { BrandContext } from '../../brand/context/BrandContext';
